refactor(corelib): Rename to fn_x
It was confusing when searching `Fn1` and the answer is `lambda.rs`
